Rename output to logger
